Struct riddle_platform_winit::Window

pub struct Window { /* fields omitted */ }

A platform native window.

Construct windows using WindowBuilder.

The window is kept open for the lifetime of this object, once it is dropped the window will be closed during the next event loop pump.

Example

let rdl =  RiddleLib::new()?;

// Create a window
let mut window = Some(WindowBuilder::new().build(rdl.context())?);

rdl.run(move |rdl| {
    if window.is_some() {
        // Drop all handles to the window. It will be closed before the next invocation
        // of this closure.
        window.take();
    } else {
        // The window is closed. Exit.
        rdl.quit();
    }
})

Implementations

impl Window

pub fn physical_size(&self) -> (u32, u32)

Get the size of the drawable area of the window in pixels

pub fn logical_size(&self) -> LogicalSize

Get the size of the drawable area of the window in logical units.

pub fn scale_factor(&self) -> f64

Get the scale factor of the window, based on how the window manager configures hidpi scaling

pub fn set_title(&self, title: &str)

Set the window title

pub fn subscribe_to_events(&self, sub: &EventSub<PlatformEvent>)

Attach a subscriber to the subset of platform events that relate to this window

Example

let rdl =  RiddleLib::new()?;
let subscriber: EventSub<PlatformEvent> = EventSub::new();

let window = WindowBuilder::new().build(rdl.context())?;
window.subscribe_to_events(&subscriber);

pub fn id(&self) -> WindowId

The window id which is used to identify this window in PlatformEvents